Upgrade nextra/blog theme to 2.0.0-beta.5, Restore _document.tsx
