+++ examples/blog/pages/_app.tsx
@@ -1,9 +1,29 @@
 import 'nextra-theme-blog/style.css'
 import type { AppProps } from 'next/app'
+import Head from 'next/head'
 import '../styles/main.css'
 
 const App = ({ Component, pageProps }: AppProps) => {
-  return <Component {...pageProps} />
+  return (
+    <>
+      <Head>
+        <link
+          rel="alternate"
+          type="application/rss+xml"
+          title="RSS"
+          href="/feed.xml"
+        />
+        <link
+          rel="preload"
+          href="/fonts/Inter-roman.latin.var.woff2"
+          as="font"
+          type="font/woff2"
+          crossOrigin="anonymous"
+        />
+      </Head>
+      <Component {...pageProps} />
+    </>
+  )
 }
 
 export default App
