+++ examples/blog/theme.config.js
@@ -1,42 +1,10 @@
-const meta = {
-  title: 'Next.js Blog Starter Kit',
-  description: 'Clone and deploy your own Next.js portfolio in minutes.',
-  image: 'https://assets.vercel.com/image/upload/q_auto/front/vercel/dps.png',
-}
+const YEAR = new Date().getFullYear()
 
 export default {
-  head: (
-    <>
-      <meta name="robots" content="follow, index" />
-      <meta name="description" content={meta.description} />
-      <meta property="og:site_name" content={meta.title} />
-      <meta property="og:description" content={meta.description} />
-      <meta property="og:title" content={meta.title} />
-      <meta property="og:image" content={meta.image} />
-      <meta name="twitter:card" content="summary_large_image" />
-      <meta name="twitter:site" content="@yourname" />
-      <meta name="twitter:title" content={meta.title} />
-      <meta name="twitter:description" content={meta.description} />
-      <meta name="twitter:image" content={meta.image} />
-      <link
-        rel="alternate"
-        type="application/rss+xml"
-        title="RSS"
-        href="/feed.xml"
-      />
-      <link
-        rel="preload"
-        href="/fonts/Inter-roman.latin.var.woff2"
-        as="font"
-        type="font/woff2"
-        crossOrigin="anonymous"
-      />
-    </>
-  ),
   footer: (
     <footer>
       <small>
-        <time>{new Date().getFullYear()}</time> © Your Name.
+        <time>{YEAR}</time> © Your Name.
         <a href="/feed.xml">RSS</a>
       </small>
       <style jsx>{`
